refactor: change store structure
