+++ src/app/store/auth/auth.state.ts
@@ -1,0 +1,9 @@
+import { IUser } from 'src/app/interfaces/user.interface';
+
+export interface AuthState {
+  user: IUser | null;
+}
+
+export const authInitialState: AuthState = {
+  user: null,
+};
