+++ src/app/store/auth/selectors/auth.selectors.ts
@@ -1,0 +1,4 @@
+import { createFeatureSelector } from '@ngrx/store';
+import { AuthState } from '../auth.state';
+
+export const selectAuth = createFeatureSelector<AuthState>('auth');
